requirements.txt
numpy==2.3.2
torch==2.4.1
python-chess==1.999
#tqdm==4.66.4
#matplotlib==3.9.2
#seaborn==0.13.2
#scikit-learn==1.5.2